Add files via upload
I was able to demonstrate my skills in prompt engineering to generate content for my topic of choice “Innovative use of sensors and actuators”  by using different LLMs such as chatGPT, POE AI, Microsoft Co-Pilot, and Gemini. So as to make the best prompts that would give me the best outputs, I used prompt tools to optimize my prompts. In comparison to my personal prompts, the optimized prompts were more detailed and thus helped me achieve the best outputs and engaging e-learning content. Furthermore, after receiving outputs from the optimized prompts, I used the outputs to further optimize my prompts and used the revised prompts to get a more clear, less ambiguous and engaging content. From this outputs, I was able to make a choice of the best content generated that is more detailed and created presentations for it. For the presentation, I used Microsoft Co-Pilot to help create visual learning aids such as pictures and graphs to help learners in understanding the content. Microsoft Co-Pilot produced the best pictures from the prompts.